Update contacts.py
For Highrise, the adresses are divided on multiple columns. The logic implemented in import_from_csv needs to be reviewed.

One other point: the method "import_from_csv" is badly implemented. It consumes a lot of time. It's better to create a contact in memory (that has the same schema of the contact model), and save it one time on Datastore and Google Drive.

A lot of work needs to be done on this part, hence, a contacts parser for CSV needs to be implemented independently.

Former-commit-id: 86fe0f03c2358da10636296b4c234c6caa404058 [formerly aa6b48dc46d168d9539b6fb28e816ca64f5f8237] [formerly 7fbf692f12456be7c862aa4efd383354b5c056ef [formerly 1c55b2e11487be2f22cd97492f50121ca491b552]]
Former-commit-id: 920be7102cd5113517c3280840ed0e0a98921d9d [formerly 17835098c25bf29de9b23ff652658a2c37eacd3d]
Former-commit-id: 9e320f5c95c5a707067c273cba2c7d93e212d2e5